requirements.txt
click
colorama
fasteners
giturlparse.py
marshmallow
pbr
PyYAML
sh
